Refactoring and tidying up the survey runner code
