+++ .travis.yml
@@ -5,6 +5,5 @@ python:
 install: "pip install -r requirements.txt"
 # command to run tests
 script:
-  - python questionnaireManagerTest.py
   - python tests.py
