+++ branching/skip.py
@@ -1,0 +1,6 @@
+from branching.condition import Condition
+
+
+class SkipCondition(Condition):
+    def __init__(self, trigger, state):
+        super(SkipCondition, self).__init__(trigger, state)
