+++ questions/inputnumber.py
@@ -1,0 +1,6 @@
+from questions.question import Question
+
+
+class InputNumber(Question):
+    def __init__(self, question_schema, parent=None):
+        super(InputNumber, self).__init__(question_schema, parent)
