+++ questions/inputtext.py
@@ -1,0 +1,6 @@
+from questions.question import Question
+
+
+class InputText(Question):
+    def __init__(self, question_schema, parent=None):
+        super(InputText, self).__init__(question_schema, parent)
